+++ app/Http/Controllers/Admin/UserRegisterController.php
@@ -26,9 +26,9 @@ class UserRegisterController extends Controller
         $request->validate([
             'name' => ['required', 'string'],
             'sex_type' => ['required', 'string'],
-            'height' => ['required', 'integer', 'max:999'],
-            'weight' => ['required', 'integer', 'max:999'],
-            'fat_percentage' => ['required', 'integer', 'max:99'],
+            'height' => ['required', 'numeric', 'max:999'],
+            'weight' => ['required', 'numeric', 'max:999'],
+            'fat_percentage' => ['required', 'numeric', 'max:99'],
             'sport_name' => ['required', 'string'],
             'sport_position' => ['string'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
